Use MinGRUBlock in MinGRUSynthetic
